deleted editorOpenWith
