+++ src/vs/workbench/contrib/files/browser/editors/binaryFileEditor.ts
@@ -13,8 +13,6 @@ import { BINARY_FILE_EDITOR_ID } from 'vs/workbench/contrib/files/common/files';
 import { IStorageService } from 'vs/platform/storage/common/storage';
 import { IWorkbenchEnvironmentService } from 'vs/workbench/services/environment/common/environmentService';
 import { IOpenerService } from 'vs/platform/opener/common/opener';
-import { openEditorWith } from 'vs/workbench/services/editor/common/editorOpenWith';
-import { IInstantiationService } from 'vs/platform/instantiation/common/instantiation';
 
 /**
  * An implementation of editor for binary files that cannot be displayed.
@@ -27,7 +25,6 @@ export class BinaryFileEditor extends BaseBinaryResourceEditor {
 		@ITelemetryService telemetryService: ITelemetryService,
 		@IThemeService themeService: IThemeService,
 		@IOpenerService private readonly openerService: IOpenerService,
-		@IInstantiationService private readonly instantiationService: IInstantiationService,
 		@IStorageService storageService: IStorageService,
 		@IWorkbenchEnvironmentService environmentService: IWorkbenchEnvironmentService,
 	) {
@@ -51,7 +48,7 @@ export class BinaryFileEditor extends BaseBinaryResourceEditor {
 			input.setForceOpenAsText();
 
 			// If more editors are installed that can handle this input, show a picker
-			await this.instantiationService.invokeFunction(openEditorWith, input, undefined, options, this.group);
+			await this.group.openEditorWith(input, undefined, options);
 		}
 	}
